Change over to storing Session pointers instead of Connection pointers
This commit includes a bunch of test output changes. This is because of how Sessions
are stored in the map. We're storing them by keyed by a hash of the key stored by the
Session objects and are therefore in the map in a different order. When the sessions
are drained at shutdown, they're drained and logged in a different order than before.